pyproject.toml
[tool.poetry]
name = "health-care-autofiller"
version = "0.1.0"
description = ""
authors = ["Ian <exor2008@yandex.ru>"]
readme = "README.md"
packages = [{include = "health_care_autofiller"}]

[tool.poetry.dependencies]
python = "^3.11"
openpyxl = "^3.1.2"
python-dotenv = "^1.0.0"
pandas = "^2.1.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
mypy = "^1.5.1"
black = "^23.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
